+++ rogue-thi-app/lib/backend-utils/timetable-utils.js
@@ -1,9 +1,10 @@
 import API from '../backend/authenticated-api'
-import LegacyAPI from '../backend/authenticated-legacy-api'
 import { combineDateTime } from '../date-utils'
 import { getNextValidDate } from './rooms-utils'
 
-import { UseLegacyAPI, setUseLegacyAPI } from './use-legacy'
+import { getFriendlyLegacyTimetable } from '../backend-legacy-utils/timetable-legacy-utils'
+
+const LEGACY_MODE = process.env.NEXT_PUBLIC_LEGACY_MODE === 'true' || false
 
 /**
  * Extracts regular, short and full names for a lecture.
@@ -70,49 +71,6 @@ export function getTimetableGaps (timetable) {
   return gaps
 }
 
-async function getFriendlyLegacyTimetable (date, detailed) {
-  const { timetable } = await LegacyAPI.getTimetable(date, detailed)
-
-  return timetable
-    .map(x => {
-      // parse dates
-      x.startDate = new Date(`${x.datum}T${x.von}`)
-      x.endDate = new Date(`${x.datum}T${x.bis}`)
-
-      // normalize room order
-      if (x.raum) {
-        x.rooms = x.raum
-          .split(',')
-          .map(x => x.trim().toUpperCase())
-          .sort()
-        x.raum = x.rooms.join(', ')
-      } else {
-        x.rooms = []
-        x.raum = ''
-      }
-
-      return {
-        date: x.datum,
-        startDate: x.startDate,
-        endDate: x.endDate,
-        name: x.fach,
-        shortName: x.veranstaltung,
-        rooms: x.rooms,
-        lecturer: x.dozent,
-        exam: x.pruefung,
-        course: x.stg,
-        studyGroup: x.stgru,
-        sws: x.sws,
-        ects: x.ectspoints,
-        objective: x.ziel,
-        contents: x.inhalt,
-        literature: x.literatur
-      }
-    })
-    .filter(x => x.endDate > date)
-    .sort((a, b) => a.startDate - b.startDate)
-}
-
 /**
  * Fetches and parses timetable data.
  * @param {Date} date Date to fetch
@@ -120,12 +78,8 @@ async function getFriendlyLegacyTimetable (date, detailed) {
  * @returns {object[]}
  */
 export async function getFriendlyTimetable (date, detailed) {
-  setUseLegacyAPI(true)
-  const legacy = UseLegacyAPI()
-
-  if (legacy) {
-    console.warn('Using legacy timetable API!')
-    return getFriendlyLegacyTimetable(date, detailed)
+  if (LEGACY_MODE) {
+    return await getFriendlyLegacyTimetable(date, detailed)
   }
 
   const { timetable } = await API.getTimetable(date, detailed)
