+++ rogue-thi-app/pages/library.jsx
@@ -10,6 +10,7 @@ import ReactPlaceholder from 'react-placeholder'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faTrashAlt } from '@fortawesome/free-solid-svg-icons'
 
+import Alert from 'react-bootstrap/Alert'
 import AppBody from '../components/page/AppBody'
 import AppContainer from '../components/page/AppContainer'
 import AppNavbar from '../components/page/AppNavbar'
@@ -25,6 +26,8 @@ import styles from '../styles/Library.module.css'
 import { Trans, useTranslation } from 'next-i18next'
 
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
+
+const LEGACY_MODE = process.env.NEXT_PUBLIC_LEGACY_MODE === 'true' || false
 
 /**
  * Page for reserving library seats.
@@ -43,6 +46,10 @@ export default function Library () {
    * Fetches and displays the reservation data.
    */
   async function refreshData () {
+    if (LEGACY_MODE) {
+      return
+    }
+
     const available = await getFriendlyAvailableLibrarySeats()
     setAvailable(available)
 
@@ -119,131 +126,144 @@ export default function Library () {
     return getAvailableRooms()[0][0][0]
   }
 
+  function LegacyWarning () {
+    return (
+      <Alert variant="danger" className={styles.alert}>
+        {t('library.legacyWarning')}
+      </Alert>
+    )
+  }
+
+  function LibraryAppBody () {
+    return (<>
+      <Modal show={!!reservationDay && !!reservationTime} onHide={hideReservationModal} onShow={() => setReservationRoom(getFirstAvailableRoom())}>
+        <Modal.Header closeButton>
+          <Modal.Title>{t('library.modal.title')}</Modal.Title>
+        </Modal.Header>
+        <Modal.Body>
+          {t('library.modal.details.day')}: {reservationDay && formatFriendlyDate(reservationDay.date)}<br />
+          {t('library.modal.details.start')}: {reservationTime && formatFriendlyTime(reservationTime.from)}<br />
+          {t('library.modal.details.end')}: {reservationTime && formatFriendlyTime(reservationTime.to)}<br />
+          <br />
+          <Form.Group>
+            <Form.Label>{t('library.modal.details.location')}:</Form.Label>
+            <Form.Control as="select" onChange={event => setReservationRoom(event.target.value)}>
+              {reservationTime && getAvailableRooms().map(([roomId, room, idx]) =>
+                <option key={idx} value={roomId}>
+                  {room.room_name}
+                </option>
+              )}
+            </Form.Control>
+          </Form.Group>
+          <Form.Group>
+            <Form.Label>{t('library.modal.details.seat')}:</Form.Label>
+            <Form.Control as="select" onChange={event => setReservationSeat(event.target.value)}>
+              <option value={-1}>{t('library.modal.seatSelection.any')}</option>
+            {reservationTime && reservationRoom &&
+              Object.values(reservationTime.resources[reservationRoom].seats).map((x, idx) =>
+              <option key={idx} value={x}>
+                {x}
+              </option>
+              )
+            }
+            </Form.Control>
+          </Form.Group>
+        </Modal.Body>
+        <Modal.Footer>
+          <Button variant="primary" onClick={addReservation}>
+            {t('library.modal.actions.reserve')}
+          </Button>
+          <Button variant="secondary" onClick={hideReservationModal}>
+            {t('library.modal.actions.cancel')}
+          </Button>
+        </Modal.Footer>
+      </Modal>
+
+      <h4 className={styles.heading}>
+        {t('library.yourReservations')}
+      </h4>
+      <ReactPlaceholder type="text" rows={3} ready={reservations}>
+        <ListGroup>
+          {reservations && reservations.length === 0 &&
+            <ListGroup.Item>
+              {t('library.details.noReservations')}
+            </ListGroup.Item>
+          }
+          {reservations && reservations.map((x, i) =>
+            <ListGroup.Item key={i}>
+              <div className={styles.floatRight}>
+                <Button variant="danger" onClick={() => deleteReservation(x.reservation_id)}>
+                  <FontAwesomeIcon title={t('library.actions.delete')} icon={faTrashAlt} />
+                </Button>
+              </div>
+
+              <Trans
+                i18nKey="library.details.reservationDetails"
+                ns='library'
+                values={{
+                  category: x.rcategory,
+                  seat: x.resource,
+                  reservation_id: x.reservation_id
+                }}
+                components={{
+                  strong: <strong />
+                }}
+              />
+              <br />
+              {formatNearDate(x.start)}: {formatFriendlyTime(x.start)} - {formatFriendlyTime(x.end)}
+            </ListGroup.Item>
+          )}
+        </ListGroup>
+      </ReactPlaceholder>
+
+      <h4 className={styles.heading}>
+        {t('library.availableSeats')}
+      </h4>
+      <ReactPlaceholder type="text" rows={20} ready={available}>
+        <ListGroup>
+          {available && available.map((day, i) =>
+            day.resource.map((time, j) =>
+              <ListGroup.Item key={i + '-' + j}>
+                {Object.values(time.resources).reduce((acc, room) => acc + room.num_seats, 0) > 0 &&
+                  <Button variant="outline-secondary" className={styles.floatRight} onClick={() => {
+                    setReservationDay(day)
+                    setReservationTime(time)
+                  }}>
+                    {t('library.actions.reserve')}
+                  </Button>
+                }
+
+                {formatNearDate(time.from)}
+                {', '}
+                {formatFriendlyTime(time.from)}
+                {' - '}
+                {formatFriendlyTime(time.to)}
+                <br />
+                <div className="text-muted">
+                  {t('library.details.seatsAvailable', {
+                    available: Object.values(time.resources).reduce((acc, room) => acc + room.num_seats, 0),
+                    total: Object.values(time.resources).reduce((acc, room) => acc + room.maxnum_seats, 0)
+                  })}
+                </div>
+              </ListGroup.Item>
+            )
+          )}
+          {available && available.length === 0 &&
+            <ListGroup.Item>
+              {t('library.details.noMoreReservations')}
+            </ListGroup.Item>
+          }
+        </ListGroup>
+      </ReactPlaceholder>
+    </>)
+  }
+
   return (
     <AppContainer>
       <AppNavbar title={t('library.title')} />
 
       <AppBody>
-        <Modal show={!!reservationDay && !!reservationTime} onHide={hideReservationModal} onShow={() => setReservationRoom(getFirstAvailableRoom())}>
-          <Modal.Header closeButton>
-            <Modal.Title>{t('library.modal.title')}</Modal.Title>
-          </Modal.Header>
-          <Modal.Body>
-            {t('library.modal.details.day')}: {reservationDay && formatFriendlyDate(reservationDay.date)}<br />
-            {t('library.modal.details.start')}: {reservationTime && formatFriendlyTime(reservationTime.from)}<br />
-            {t('library.modal.details.end')}: {reservationTime && formatFriendlyTime(reservationTime.to)}<br />
-            <br />
-            <Form.Group>
-              <Form.Label>{t('library.modal.details.location')}:</Form.Label>
-              <Form.Control as="select" onChange={event => setReservationRoom(event.target.value)}>
-                {reservationTime && getAvailableRooms().map(([roomId, room, idx]) =>
-                  <option key={idx} value={roomId}>
-                    {room.room_name}
-                  </option>
-                )}
-              </Form.Control>
-            </Form.Group>
-            <Form.Group>
-              <Form.Label>{t('library.modal.details.seat')}:</Form.Label>
-              <Form.Control as="select" onChange={event => setReservationSeat(event.target.value)}>
-                <option value={-1}>{t('library.modal.seatSelection.any')}</option>
-              {reservationTime && reservationRoom &&
-                Object.values(reservationTime.resources[reservationRoom].seats).map((x, idx) =>
-                <option key={idx} value={x}>
-                  {x}
-                </option>
-                )
-              }
-              </Form.Control>
-            </Form.Group>
-          </Modal.Body>
-          <Modal.Footer>
-            <Button variant="primary" onClick={addReservation}>
-              {t('library.modal.actions.reserve')}
-            </Button>
-            <Button variant="secondary" onClick={hideReservationModal}>
-              {t('library.modal.actions.cancel')}
-            </Button>
-          </Modal.Footer>
-        </Modal>
-
-        <h4 className={styles.heading}>
-          {t('library.yourReservations')}
-        </h4>
-        <ReactPlaceholder type="text" rows={3} ready={reservations}>
-          <ListGroup>
-            {reservations && reservations.length === 0 &&
-              <ListGroup.Item>
-                {t('library.details.noReservations')}
-              </ListGroup.Item>
-            }
-            {reservations && reservations.map((x, i) =>
-              <ListGroup.Item key={i}>
-                <div className={styles.floatRight}>
-                  <Button variant="danger" onClick={() => deleteReservation(x.reservation_id)}>
-                    <FontAwesomeIcon title={t('library.actions.delete')} icon={faTrashAlt} />
-                  </Button>
-                </div>
-
-                <Trans
-                  i18nKey="library.details.reservationDetails"
-                  ns='library'
-                  values={{
-                    category: x.rcategory,
-                    seat: x.resource,
-                    reservation_id: x.reservation_id
-                  }}
-                  components={{
-                    strong: <strong />
-                  }}
-                />
-                <br />
-                {formatNearDate(x.start)}: {formatFriendlyTime(x.start)} - {formatFriendlyTime(x.end)}
-              </ListGroup.Item>
-            )}
-          </ListGroup>
-        </ReactPlaceholder>
-
-        <h4 className={styles.heading}>
-          {t('library.availableSeats')}
-        </h4>
-        <ReactPlaceholder type="text" rows={20} ready={available}>
-          <ListGroup>
-            {available && available.map((day, i) =>
-              day.resource.map((time, j) =>
-                <ListGroup.Item key={i + '-' + j}>
-                  {Object.values(time.resources).reduce((acc, room) => acc + room.num_seats, 0) > 0 &&
-                    <Button variant="outline-secondary" className={styles.floatRight} onClick={() => {
-                      setReservationDay(day)
-                      setReservationTime(time)
-                    }}>
-                      {t('library.actions.reserve')}
-                    </Button>
-                  }
-
-                  {formatNearDate(time.from)}
-                  {', '}
-                  {formatFriendlyTime(time.from)}
-                  {' - '}
-                  {formatFriendlyTime(time.to)}
-                  <br />
-                  <div className="text-muted">
-                    {t('library.details.seatsAvailable', {
-                      available: Object.values(time.resources).reduce((acc, room) => acc + room.num_seats, 0),
-                      total: Object.values(time.resources).reduce((acc, room) => acc + room.maxnum_seats, 0)
-                    })}
-                  </div>
-                </ListGroup.Item>
-              )
-            )}
-            {available && available.length === 0 &&
-              <ListGroup.Item>
-                {t('library.details.noMoreReservations')}
-              </ListGroup.Item>
-            }
-          </ListGroup>
-        </ReactPlaceholder>
-
+        {LEGACY_MODE ? <LegacyWarning /> : <LibraryAppBody />}
         <AppTabbar />
       </AppBody>
     </AppContainer>
